Feat: 트랙 팀전 기록 추가
- README 수정
- interface 코드 모두 한 곳으로 이동
- 트랙 테이블 tr 컴포넌트 분리 및 렌더링 최적화
