+++ README.md
@@ -16,6 +16,12 @@ PROJECT PERIOD: 2022.03.21 ~ 2022.03.25
 
 <br>
 
+### 참고
+
+로컬 실행시 firebase 설정 값들을 환경변수 처리하여 응원 한마디 부분이 제대로 저장되지 않습니다.
+
+<br>
+
 ## ✨ 주요 기능
 
 - 홈 페이지
@@ -28,15 +34,11 @@ PROJECT PERIOD: 2022.03.21 ~ 2022.03.25
   - 트랙/카트 탭에서 내가 많이 플레이한 트랙/카트를 알 수 있습니다.
   - 최근 200경기의 매치 기록과 같이 플레이한 유저들의 기록을 확인할 수 있습니다. (50경기씩 무한 스크롤)
 - 트랙 페이지
-  - 오늘 하루 동안 스피드 개인전 트랙 순위를 볼 수 있습니다.
+  - 오늘 하루 동안의 스피드 개인전 / 팀전 트랙 순위를 볼 수 있습니다.
   - 지표 가이드로 지표가 나타내는 바를 알 수 있습니다.
   - 테이블의 지표를 통해 트랙별 베스트 라이더, 베스트 레코드 등을 확인할 수 있습니다. (30 트랙씩 무한 스크롤)
 - 랭킹, 카트 페이지
   - 추후 서비될 예정인 페이지 공지를 하였습니다.
-
-## 참고
-
-로컬 실행시 firebase 설정 값들을 환경변수 처리하여 응원 한마디 부분이 제대로 저장되지 않습니다.
 
 <br>
 
@@ -58,7 +60,7 @@ PROJECT PERIOD: 2022.03.21 ~ 2022.03.25
 
 3. 트랙 페이지
 
-<img src="https://user-images.githubusercontent.com/36434219/160115059-f819ce9d-1b2b-4265-baaa-9185ffcb9419.png" alt="트랙 페이지" width="700px" height="400px">
+<img src="https://user-images.githubusercontent.com/36434219/160732220-f987a4e2-d600-4860-bee1-9cc5b620230c.png" alt="트랙 페이지" width="700px" height="400px">
 
 <br>
