+++ src/components/common/Footer.tsx
@@ -1,5 +1,5 @@
-import styled from '@emotion/styled'
 import React from 'react'
+import styled from '@emotion/styled'
 
 export default function Footer() {
   return (
